Temporary disable SA assertions for port list with limit
An API request for a portlisting filtered by ip-address or subnet
failed with a 500, as get_collection_query() in _get_ports_query()
already sets a limit on the query. Subsequent query.filter() calls
fail with an exception. To circumvent this we disable the SA
assertions around this query, including _no_limit_offset() until
a real fix is available. We can do this because the assertion is
in place to prevent someone to apply a filter() on a already limit()ed
query, as the filter will be executed first in any case - but this is
exactly what we want here.

The bug was introduced by Openstack change-id Ia6e916a9dae5f5ef203259c3d1c6e1c30445c07d
and commit 5b7c8c3e27dab38a3b336bb3ac0a87b98944b6cf