Revert get_active_networks pagination in dhcp agent
In our special usecase pagination of this function did not yield a
high enough performance increase, as it also caused agent network scheduling
to be triggered multiple times per sync loop. With the other optimizations
we do not need this pagination anymore, as we do not run into timeouts for
this particular call.

This reverts the following commits:
 * 5ef758b3cb4fc1b0c733189577ceec8cc0999395 pagination for dhcp agent get_active_networks
 * 5a88f0f78337e6bdbeacbafab71d243570a82079 fix index out of range for dhcp sync_state
 * 9556ff348b02911935af57407a2810bdff00dd00 externd, not append!