Add explicit policy-enforcement for port.security_groups
Our customers want to be able to prohibit certain colleagues from
changing the security_groups. For this, we have to tell neutron to
explicitly enforce policy for `Port.security_groups`.
After that, neutron checks for `update_port:security_groups` and
`create_port:security_groups` in addition to `update_port` and
`create_port`.