get_detailed_tenant_quotas: Check for existing resource before accessing it from dict
Upgrading Neutron from lbaasv2 to non-lbaasv2 breaks quota detail if theres been already
quotas set for example 'healthmonitor'. This is caused becaused the quota table has
been already populated with (now deprecated) lbaasv2 quota limits which are unconditionally
used to populate the quota detail dictionary result - which is missing this key.

This commit add a check for the existence of the key before accessing the dictionary.