Restrict query of iface attrs in linuxbridge agent
For a linuxbridge-agent the startup performance gets worse the more
networks are scheduled onto the agent. One factor in this is that
parsing and transporting all netlink attributes for all interfaces takes
a long time. This makes functions like get_link_devices() pretty slow.
The more interfaces there are the slower it gets.

To prevent this an attribute filter is introduced to get_link_devices()
and to make_serializable(). Only specified attributes are serialized for
transport from each interface. As attributes can also be nested, this
can save up a lot of processing time, especially when only the interface
name is needed by the caller.

get_devices_info() now allows interface attributes to be restricted to a
subset as well. The function will determine which netlink attributes
need to be requested.

Co-authored-by: Johannes Kulik <johannes.kulik@sap.com>

Fix wrong fields in get_devices_info() for linuxbridge

A missing ',' merged two fields in get_devices_info(), causing the
fields to be missing from the resulting dictionary.