Add address scope to router sync data
Upstream Neutron has the assumption that all subnets of a network are in
the same subnet pool and uses this assumption to create some shortcuts.
In this case the address scope reported for a router port is reported as
the address scope of the last subnet in the network (independant of if
the router port is in that subnet).

To fix this - and let the consumer of this data decide what to do with
this information - we add the address scope of a subnet to each port's
subnet list.