Only set address scope when present for l3 sync data
Upstream Neutron has the assumption that all subnets of a network are in
the same subnet pool (and therefore same address scope), while our brand
of Neutron requires all subnets in a network to be either in the same
address scope or in no address scope at all.

Neutron adds the address scope to each port in the l3 router sync data,
but looks at all subnets present in the network. We now change this
behavior to only add the address scope when the router port is actually
in a subnet with this address scope. With this DAPNets (Directly
Accessible Private Networks) and non-DAPNets can live in peace together.