Relax check on subnet pools to just satisfy same address scope
The current implementation does not allow subnets from different address
scopes in a network. For DAPnets we need to relax that constraint so we
allow subnets from the same address scope as well as subnets with no
subnet pool (and hence no address scope) and subnets with the same
address scope in a network.