Mac address must only be globally unique on create
The Cisco IOS XE platform drops any frames with a mac address that it has bound
on a local interface, no matter which VRF they are in. OpenStack generates mac
addresses randomly and checks only if the address is already used in the
current network. We implemented global mac address duplicate checking in
1a58780b81, but had to notice that ironic baremetal deployments may generate 2
ports in different subnets with the same mac. However this is only done on an
update and a mac address update is an admin-only action in our cloud. Hence we
relax the check for update operations to only be network-wise unique while the
create check must satisfy a globally unique constraint.