Parsable logging for metadata service requests
We want to get a better impression of what our users do with our
metadata service and for this we want to improve the logging. The
current logging logs part of the request (part by eventlet.wsgi, part by
Neutron itself) or the complete ports request ("which ports do belong to
the requester) as a multiline string representation of the request
object.

To make debugging this easier for us we now log an extra single line for
each request, containing many informations, as the current OpenStack
network, client ip, user-agent, request path etc., which can then be
easily consumed by some log parser (fluentd + grok in our case).