Prefer default gw with ip on port on DHCP Agent
When multiple subnets are configured in the same network, the DHCP agent
will choose the first suitable subnet with a gateway ip set as default
gateway, where the subnets are in a generally stable, but arbitrary
order. In some cases only some of the subnets in a network actually have
a router in it, meaning a DHCP agent might loose Internet access (used
for example for DNS).

Now we sort the subnets before selecting a default gateway. Subnets are
sorted first by if their gateway ip is present on a port in the network
and then by their created_at time, so we always have a stable order.
Also, using the first created subnet seems like a good decision, as it
will probably be the one with the least fluctuations. We won't filter
for a router port, as the user could also provide a custom router via a
VM in their network. In case no subnet is present on a port we will
still use one of them as default route, just to preserve the old
behavior.