Prevent ext subnet gw IP being allocated as fips
External subnets might have an external gateway IP outside of the
subnets allocation pool. This IP is not allocated by the OpenStack IPAM,
but is used as default gateway. A user might create a floating IP for
this specific IP, which results in double usage of the IP. To prevent
this we disallow allocating gateway IPs of external subnets as floating
IPs.

This commit only checks that on FIP creation the IP is not a gateway IP
of any subnet it is allocated inside. We do not check this condition if
the subnet's gateway IP is modified.

This bug has been discussed with upstream, but was found to be invalid
by them, because in the default configuration only admins can create a
FIP with a predetermined IP address. For us this is different, but
additionally I think we also shouldn't allow admins to do this, as in
our infra this brings the network down.

Discussion with upstream can be found here:
https://bugs.launchpad.net/neutron/+bug/1959699