Proper iface deletion in agent with pyroute 0.6.11
With pyroute2 0.6.6 and a Linux Kernel since 5.19, pyroute2 sometimes
cannot properly delete interfaces. This results in the cleanup process
being interrupted, i.e. dangling vlan bond interfaces and dangling
bridges. Especially the bonds can result in a bond that cannot be
properly put into a bridge at some point, which makes vlan reuse on an
agent impossible.

To solve this problem we need to bump pyroute2 to 0.6.11, where this has
been fixed[0][1]. The changelog of pyroute2[2] looks like we can safely
upgrade and a manual upgrade in one network agent looks also as if this
is working.

The Kernel patch introducing the problem can be found here[3], OpenStack
bug here[4] and Ubuntu package bug here[5].

[0] https://github.com/svinota/pyroute2/issues/892
[1] https://github.com/svinota/pyroute2/issues/923
[2] https://docs.pyroute2.org/changelog.html
[3] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/diff/?id=a6cec0bcd34264be8887791594be793b3f12719f
[4] https://bugs.launchpad.net/ubuntu/+source/neutron/+bug/1995735
[5] https://bugs.launchpad.net/ubuntu/+source/pyroute2/+bug/1995469