Switch to our own oslo.log
Our own oslo.log is based on 4.7.0 (latest oslo.logyoga release) with an
additional patch to prevent deadlocks when logging is used with native
and greenthreads ad the same time.

This is only needed for Neutron yoga and should be dropped on a later
version of Neutron. The version, even with containing some extra commits
is hard-set on 4.7.0 to work with the current upper-constraints.