customdns - fix keystone session, needed for caracal/sdk update
After updating to Caracal and using openstacksdk 3.0.0 the sdk
refused initializing the identity client correctly to talk to
keystone.

This apparently happens when a full config is given to the Connection,
and should not have worked before.

Also now calling identity directly without indirection, which also
should be more efficient because it uses a direct get for the id.