[OVN] don't set explicit down status for ports that just got created
OVN expects more or less to be the only driver responsible for ports.

This patch removes the explicit port_status_down for a newly created
logical_switch_port in OVSDB-NB. OVN driver will sync all network ports
to the ovsdb, regardless if they are bound against phyiscal network with
ovn-controller or not. So this patch leaves the status intact for ports
that are not managed by OVN.

status=false is anyway the default for new ports.