[OVN] use empty string for OVN schema when vnic_type is not defined.
We encountered neutron port's with **None** `vnic_type`, which cannot
modeled in OVSDB, thus need to be converted to a empty string.

When a port only has binding profiles with status INACTIVE it looks like
this gets returned as the port not having a vnic_type. This is something
that should not happen, but still can happen due to botched migrations.
Therefore we need this workaround until we can make sure we always have
a binding in status ACTIVE for our ports.